Fix up \todo notes in float captions.
The default margin placement doesn't work in float contexts; this is
listed as a known issue in Section 1.6.9 of the todonotes documentation
(https://www.ctan.org/pkg/todonotes). \todo[inline] works, but you have
to position it *outside* the \caption{} and *after* the \label{},
otherwise the \label{} attaches to the todo note, and references to the
figure/section will instead refer to the following(?) section.