Define and use a \firstterm macro.
This is meant for the first introduction of a term that will be reused
later, to highlight it as a meaningful phrase in itself and not generic
language. I had been using \emph for this, but I want something
semantically distinct. According to
https://www.overleaf.com/learn/latex/Bold%2C_italics_and_underlining#Emphasising_text,
the behavior of \emph may change in some contexts.