Simplify estimation of snowflake-02 ratio.
Start with userstats-bridge-transport-multi.csv rather than
userstats-bridge-combined-multi.csv as we don't need country
information. The numbers are pretty close to what they were before,
except that the snowflake-01 count had been weirdly high, maybe
something to do with the low/high estimation on a day when the bridge
was restarted.

 %    date       transport `snowflake-01` `snowflake-02`
-%  3 2023-02-17 snowflake        120075.          3222.
+%  3 2023-02-17 snowflake         70305.          3221.