Nix -tolerance from proxy-churn-windows.go.
On second thought, I don't want windows to be interrupted by gaps in
coverage (which happens when the broker process is restarted). I want to
see how the decay goes, even across a restart.

The little bit of wiggle room that -tolerance formerly provided when
deciding what input records to include in a window, we can get by
slightly expanding -duration, -before, and -after.