Update notes about DTLS client and server roles.
The observation that the Snowflake client is not always the DTLS client
only matters for the supported_groups fingerprint in Client Hello, which
we are not going into. For the presence of Hello Verify Request, which
is sent by the DTLS server, the situation is the same as with
supported_groups in Server Hello: a non-Pion in the server role evades
it.