Compensate for days when coverage < num_instances.
When we run N tor instances per bridge, we expect N bridge-extra-info
descriptors to be published each day. But sometimes there are fewer than
N descriptors published. This happens whenever the bridge is restarted
(which restarts the 24-hour clock for publishing a descriptor), but it
also happens that on some days, not all the instances publish a
descriptor, for reasons unknown. When this happens, we count fewer users
than we should, because the available descriptors cover only part of a
day, or the whole day is covered but only by a subset of descriptors.

This change extrapolates the counts, as if we were not missing any
descriptor coverage. If we see only m/N descriptors in a day, we
multiply the observed values by N/m. In actual fact, we multiply by
N/max(m, N), because there are times, such as when we were testing or
migrating between staging servers, when there are more descriptors
published than the nominal number of instances.

Compare to
https://gitlab.torproject.org/dcf/snowflake-graphs/-/commit/57dabb4861d62acdbd00cfe42f62d872955ed5f2

Although the proxy CSVs have a coverage columns, I have decided not to
scale the proxy graphs (proxy-type.pdf, proxy-nat-type.pdf). Unique IP
addresses do not scale linearly with time like the other partial
quantities do. I tried it, and it resulted in spikes that were clearly
unnatural.