Estimate and subtract non-Snowflake (<OR>) bytes from bandwidth.
The bandwidth information in descriptors does not have a transport
attached; but we can estimate bandwidth by transport by assuming that
the ratio is the same as for user counts. This is hardly worth doing, as
the non-Snowflake users are less than 1% of users. (Honestly I'm not
sure where those non-Snowflake counts come from; I've always assumed it
was some internal Tor connections or something.)

Unfortunately it results in a gap in the bandwidth gap that was not
there before, in May–June 2021, because we lack user counts during that
time and so there is no estimate of the ratio of Snowflake users to
non-Snowflake users.