Cut sec:challenges.
This section was discursive, not focused on any reader need. I had
intended this section originally mainly to impress upon the reader the
extreme difficulty of working with libwebrtc which had the project
stalled for such a long time and formed such a large part of its
history. But it's hard to convey an accurate impression to the reader,
and unclear that there is even value in doing so. Me venting my
frustration is arguably not a good look. We already make the point of
transitioning from Chromium to Pion, in a softer manner, in
sec:fingerprinting and sec:deployment. Researching and documenting the
history of the transition was privately cathartic, but that's as far as
it needs to go.

The only other real substance in the section was the observation that
some users may be confused about whether they need to run the browser
extension and/or the Orbot proxy. I was already thinking of cutting that
paragraph for being uninteresting, and without it, there's not much to
support the subsection.