Avoid the implication that "no-secrets" is a necessary feature of rendezvous.
There is disagreement among the authors about whether the absence of
secret information is a necessary feature of rendezvous-based systems,
as the text formerly implied, or if other systems with a multi-tiered
connection process might also be included under that heading, even if
they involve out-of-band secrets. To avoid taking a position on that
point, just say that no-secrets is a property of the systems
specifically listed.