Adjust footnote formatting.
The main point of this is to keep the second line of the footnote at the
same level of indentation as the first; otherwise it looks bad with URLs
that are split across lines.

   9 https://bugs.torproject.org/tpo/anti-censorship/pluggable-transports/snowflake/
30931#note_2593598
  10 https://bugs.torproject.org/tpo/anti-censorship/pluggable-transports/snowflake/
30999#note_2593718

becomes

 9 https://bugs.torproject.org/tpo/anti-censorship/pluggable-transports/snowflake/
   30931#note_2593598
10 https://bugs.torproject.org/tpo/anti-censorship/pluggable-transports/snowflake/
   30999#note_2593718

Decreasing the overall indentation level also lets us slightly increase
the font size, while still keeping most URLs on one line.