Upgrade TorKameleon citation to TrustCom publication.
Keeping the arxiv.org URL because there is no IEEE Xplore page for it
yet, and because even when that appears it's likely to be paywalled.

Using TrustCom itself for the booktitle though it's categorized under
Intelligent and Trustworthy Computing, Communications, and Networking
(ITCCN) on the TrustCom 2023 site:
https://hpcn.exeter.ac.uk/trustcom2023/accepted-paper-list.php
because the proceedings seem to be the same. Afonso Vilalonga points to
the example of https://ieeexplore.ieee.org/document/10063644 ("The Need
for Speed"), which was in the workshop on Cyberspace Security and
Artificial Intelligence (CAI-2022), but has TrustCom as the venue from
IEEE Xplore.
https://web.archive.org/web/20221205172940/http://www.ieee-hust-ncc.org/2022/TrustCom/Workshops.html
https://web.archive.org/web/20230908122250/http://www.ieee-hust-ncc.org/2022/TrustCom/TrustCom2022%20Program%20Booklet-V6.pdf
Maybe the TrustCom workshops are not considered separate venues, but are
just a way the larger conference is subdivided.