Tweak fig:proxy-count-decay.
Adjust the phrasing to avoid the misunderstanding that we only look at
up to 24 hours later -- we actually go up to 40 hours later, but always
using windows of 24 hours.

Change the color -- saddlebrown had come to me in a dream, but this is a
better match with the proxy type graphs.